requirements.txt
opencv
pyserial
dlib
face_recognition
openai
numpy
gtts
playsound
pytest
mediapipe
